requirements.txt
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
appdirs==1.4.4
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
curl_cffi==0.13.0
distro==1.9.0
docstring_parser==0.17.0
feedparser==6.0.11
frozendict==2.4.7
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jiter==0.13.0
lxml==6.0.2
multitasking==0.0.12
numpy==2.4.2
oauthlib==3.3.1
pandas==3.0.0
peewee==3.19.0
platformdirs==4.9.2
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
webencodings==0.5.1
websockets==16.0
yfinance==1.2.0
matplotlib==3.8.2
